Improved timestep wizard to prevent instability from tracer forcing
Since the CFL number for the timestepper is ∼UΔt/Δx for velocity then it
seems reasonable that for tracer forcing it should be ∼GcΔt/C where Gc
is the tracer tendency with units [C]/s. I think I should be able to derive
this properly at some point but in the mean time I've tried it and it
seems todo a good job of dropping the timestep when a dramatic change is
occuring in tracer concentration and prevents them from becoming negative.