refactor: use variadic function for `scale_for_negs!` kernel
If list of fields to scale is passed via a tuple, in CUDA.jl it results
in a local copy of the parameter. Since it is put into the 'Local'
memory of each thread, it introduces extra memory traffic which
significantly slows down the performance of the kernel.

This problem is avoided if each field and scalefactor is passed as
parameter of a kernel directly. However to support variable number of
fields we need to make the kernel variadic and use inlining + recursion
to effectivly unroll the loops.